% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayeslmm-package.R
\docType{package}
\name{bayeslmm-package}
\alias{bayeslmm}
\alias{bayeslmm-package}
\title{bayeslmm: A Bayesian Linear Mixed Effect Model}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Ron Lewis \email{ron.lewis@dfo-mpo.gc.ca}

}
\keyword{internal}
