% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blmm.R
\name{lmm}
\alias{lmm}
\title{Perform a Bayesian linear mixed effects model}
\usage{
lmm(formula, data, cores = 4, seed = 10101)
}
\arguments{
\item{formula}{An R model formula for mixed effects models}

\item{data}{The name of a data frame}

\item{cores}{The number of CPU cores that the model will use}

\item{seed}{The initial conditions of the random number generator}
}
\value{
A list with a two element: model and waic.
}
\description{
Perform a Bayesian linear mixed effects model
}
\examples{
\dontrun{
M <- lmm(rt ~ days + (1|subject), data)
}
}
