% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blmm.R
\name{lmmplot}
\alias{lmmplot}
\title{A linear mixed effect facet plot}
\usage{
lmmplot(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
A ggplot2 plot
}
\description{
A linear mixed effect facet plot
}
\examples{
\dontrun{
lmmplot(sleepstudy)
}
}
