% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sleepstudy_df}
\alias{sleepstudy_df}
\title{A sleepstudy data frame}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 180 rows and 3 columns.
}
\usage{
sleepstudy_df
}
\description{
A sleepstudy data frame
}
\keyword{datasets}
