% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blmm.R
\name{lmm}
\alias{lmm}
\title{Perform a Bayesian linear effect model}
\usage{
lmm(formula, data, cores = 4, seed = 10101)
}
\arguments{
\item{formula}{An R model formula for mixed effect model}

\item{data}{The name of the data frame}

\item{seed}{The inizial conditions of the ranmber generator}

\item{core}{The number of CPU cores that the model will use}
}
\value{
A list with two elements: model and waic.
}
\description{
Perform a Bayesian linear effect model
}
\examples{
\dontrun{
M <- lmm(rt ~ days + (1|subject), data)
}
}
