% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blmm.R
\name{lmmplot}
\alias{lmmplot}
\title{Linear mixed effect facet plot}
\usage{
lmmplot(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
A ggplot object
}
\description{
Linear mixed effect facet plot
}
\examples{
\dontrun{
lmmplot(sleepstudy)
}
}
